requirements.txt
# Core
streamlit>=1.32.0
langchain>=0.2.0
langchain-community>=0.0.20
openai>=1.0.0

# Embeddings & Transformers
transformers>=4.40.0
sentence-transformers>=2.2.2

# Elasticsearch
elasticsearch>=8.11.0

# OCR
PyMuPDF>=1.23.7
Pillow>=10.2.0
easyocr>=1.7.1
numpy>=1.24.0

# Others
python-dotenv>=1.0.0
tqdm>=4.66.0
